+++ server/core/config.cc
@@ -137,6 +137,7 @@ const char CN_SSL_CERT_VERIFY_DEPTH[]         = "ssl_cert_verify_depth";
 const char CN_SSL_KEY[]                       = "ssl_key";
 const char CN_SSL_VERSION[]                   = "ssl_version";
 const char CN_STRIP_DB_ESC[]                  = "strip_db_esc";
+const char CN_SUBSTITUTE_VARIABLES[]          = "substitute_variables";
 const char CN_THREADS[]                       = "threads";
 const char CN_THREAD_STACK_SIZE[]             = "thread_stack_size";
 const char CN_TYPE[]                          = "type";
@@ -458,11 +459,27 @@ void fix_section_name(char *section)
  * @param value         The Parameter value
  * @return zero on error
  */
-static int
-ini_handler(void *userdata, const char *section, const char *name, const char *value)
+static int ini_handler(void *userdata, const char *section, const char *name, const char *value)
 {
-    CONFIG_CONTEXT   *cntxt = (CONFIG_CONTEXT *)userdata;
-    CONFIG_CONTEXT   *ptr = cntxt;
+    CONFIG_CONTEXT *cntxt = (CONFIG_CONTEXT *)userdata;
+    CONFIG_CONTEXT *ptr = cntxt;
+
+    if (config_get_global_options()->substitute_variables)
+    {
+        if (*value == '$')
+        {
+            char* env_value = getenv(value + 1);
+
+            if (!env_value)
+            {
+                MXS_ERROR("The environment variable %s, used as value for parameter %s "
+                          "in section %s, does not exist.", value + 1, name, section);
+                return 0;
+            }
+
+            value = env_value;
+        }
+    }
 
     if (strcmp(section, CN_GATEWAY) == 0 || strcasecmp(section, CN_MAXSCALE) == 0)
     {