+++ Documentation/Changelog.md
@@ -11,6 +11,8 @@
 * Firewall can now prevent the use of functions in conjunction with
   certain columns.
 * Parser of MaxScale extended to support window functions and CTEs.
+* Parser of MaxScale extended to support PL/SQL compatibility features
+  of upcoming 10.3 release.
 * Prepared statements are now parsed and the execution of read only
   ones will be routed to slaves.
 * Server states are persisted, so in case of crash and restart MaxScale
@@ -20,6 +22,7 @@
 * The Masking filter can now both obfuscate and partially mask columns.
 * Binlog router supports MariaDB 10 GTID at both ends.
 * KILL CONNECTION can now be used through MaxScale.
+* Environment variables can now be used in the MaxScale configuration file.
 
 For more details, please refer to:
 * [MariaDB MaxScale 2.2.0 Release Notes](Release-Notes/MaxScale-2.2.0-Release-Notes.md)